refactor: move getBrowserRenderableType to utils/fileTypes
